working on the receipt reconstruction

+++ android_app/app/src/main/java/ch/heigvd/scanalyze/Product.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Product(val id: Int,
     val abbreviatedName: String,
-    val quantity: Double,
-    val unitPrice: Double,
-    val discount: Double = 0.0): Parcelable {
+    val quantity: Float,
+    val unitPrice: Float,
+    val discount: Float = 0.0f): Parcelable {
 }
